requirements.txt
mediapipe==0.10.11
numpy==1.21.5
opencv_contrib_python==4.10.0.84
opencv_python==4.5.5.64
Pillow==10.4.0
Requests==2.32.3
streamlit==1.36.0
torch==1.10.1
torchvision==0.11.2
